+++ wordpress-bench/attributes/default.rb
@@ -1,22 +1,6 @@
-extend WordpressBench::Helpers
-
 # Cloud WorkBench
 default['wordpress-bench']['metric_name'] = 'response_time'
 default['wordpress-bench']['num_repetitions'] = 1
-
-# wp-cli
-default['wordpress-bench']['wp_cli_url'] = 'https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar'
-default['wordpress-bench']['wp_cli_bin'] = '/usr/local/bin/wp'
-
-# General
-default['wordpress-bench']['public_ip_query'] = 'wget http://ipecho.net/plain -O - -q'
-
-# Wordpress site
-default['wordpress-bench']['url'] = "http://#{default_ip}"
-default['wordpress-bench']['title'] = 'Cloud Benchmarking'
-default['wordpress-bench']['admin_user'] = 'admin'
-default['wordpress-bench']['admin_password'] = 'admin'
-default['wordpress-bench']['admin_email'] = 'admin@example.com'
 
 # Fake data generator
 default['wordpress-bench']['500px_customer_key'] = ''
@@ -25,8 +9,9 @@ default['wordpress-bench']['batch_size'] = 25
 # Load generator endpoint
 default['wordpress-bench']['load_generator'] = 'http://localhost'
 
+http_site = node['wordpress']['site']['url']
 # JMeter properties
-default['wordpress-bench']['jmeter']['properties']['site'] = default_ip
-default['wordpress-bench']['jmeter']['properties']['httpclient.timeout'] = 60 * 1000 # milliseconds
+default['wordpress-bench']['jmeter']['properties']['site'] = (http_site.sub('http://', '') rescue '127.0.0.1')
+default['wordpress-bench']['jmeter']['properties']['httpclient.timeout'] = 60 * 1000 # in milliseconds (x seconds * 1000)
 default['wordpress-bench']['jmeter']['properties']['num_threads'] = 2 # i.e., users
 default['wordpress-bench']['jmeter']['properties']['ramp_up_period'] = 0 # seconds