+++ wordpress-bench/files/default/cleanup.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+rm results.csv
+rm test_plan.jtl
+rm jmeter.log