+++ wordpress-bench/recipes/stop_services.rb
@@ -1,4 +1,6 @@
-execute 'stop-mysql-and-apache2' do
-  command 'sudo service apache2 stop; sudo service mysql-default stop'
-  action :run
+services = %w(apache2 mysql-default)
+services.each do |service_name|
+  service service_name do
+    action :stop
+  end
 end